requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2==2.9.2
pyinstaller==4.3
pylint==2.7.4
sqlparse==0.4.2
gunicorn==20.1.0
whitenoise==6.2.0